AdminFuctionality

+++ MyCarStatistics/Controllers/CarController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Identity;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using MyCarStatistics.Contracts;
 using MyCarStatistics.Data.Models.Account;
@@ -20,7 +21,12 @@ namespace MyCarStatistics.Controllers
 
         [HttpPost]
         public async Task<IActionResult> Overview(int carId)
-        {            
+        {
+            var user = await userManager.FindByNameAsync(User?.Identity?.Name);
+            if ((await carService.CheckUser(carId, user.Id.ToString())) == false)
+            {
+                return RedirectToAction("AccessDenied", "Car");
+            }
             var carOverview = await carService.GetOverviewData(carId);
             return View(carOverview);
         }
@@ -28,7 +34,12 @@ namespace MyCarStatistics.Controllers
         [HttpPost]
         public async Task<IActionResult> Edit(CarViewModel car, int carId)
         {
-            if (car.Brand == null)
+            var user = await userManager.FindByNameAsync(User?.Identity?.Name);
+            if ((await carService.CheckUser(carId, user.Id.ToString())) == false)
+            {
+                return RedirectToAction("AccessDenied", "Car");
+            }
+            if (!ModelState.IsValid)
             {
                 var entity = await carService.GetCarInfo(carId);
                 return View(entity);
@@ -41,6 +52,11 @@ namespace MyCarStatistics.Controllers
         [HttpPost]
         public async Task<IActionResult> Delete(int carId)
         {
+            var user = await userManager.FindByNameAsync(User?.Identity?.Name);
+            if ((await carService.CheckUser(carId, user.Id.ToString())) == false)
+            {
+                return RedirectToAction("AccessDenied", "Car");
+            }
             await carService.Delete(carId);
             return RedirectToAction(nameof(All));
         }
@@ -74,18 +90,14 @@ namespace MyCarStatistics.Controllers
         }
 
         [HttpPost]
+        [AllowAnonymous]
         public IActionResult Index()
-            => Ok();
+            => View();
 
+        [HttpGet]
+        public IActionResult AccessDenied()
+                => View();
 
-        private async Task<IActionResult> CheckUser(int carId)
-        {
-            ApplicationUser user = await userManager.FindByNameAsync(User?.Identity?.Name);
-
-
-
-
-            return RedirectToAction(nameof(All));
-        }
+        
     }
 }
